(refactor) cli: extract person-ID parsing/resolution utilities
Extract duplicated parsePersonIds, readPersonIdsFile, and resolution
boilerplate from 5 CLI handler files into a shared person-ids.ts module
with a new resolvePersonIds function that handles mutual exclusivity
checks, dispatch, and empty-set validation.
