+++ packages/cli/src/handlers/campaign-start.ts
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 // Copyright (C) 2025 Alexey Pelykh
 
-import { readFileSync } from "node:fs";
-
 import {
   CampaignExecutionError,
   CampaignNotFoundError,
@@ -15,34 +13,7 @@ import {
   withInstanceDatabase,
 } from "@lhremote/core";
 
-function parsePersonIds(raw: string): number[] {
-  return raw
-    .split(",")
-    .map((s) => s.trim())
-    .filter((s) => s.length > 0)
-    .map((s) => {
-      const n = Number(s);
-      if (!Number.isInteger(n) || n <= 0) {
-        throw new Error(`Invalid person ID: "${s}"`);
-      }
-      return n;
-    });
-}
-
-function readPersonIdsFile(filePath: string): number[] {
-  const content = readFileSync(filePath, "utf-8");
-  return content
-    .split(/[\n,]/)
-    .map((s) => s.trim())
-    .filter((s) => s.length > 0)
-    .map((s) => {
-      const n = Number(s);
-      if (!Number.isInteger(n) || n <= 0) {
-        throw new Error(`Invalid person ID in file: "${s}"`);
-      }
-      return n;
-    });
-}
+import { resolvePersonIds } from "./person-ids.js";
 
 /** Handle the {@link https://github.com/alexey-pelykh/lhremote#campaigns | campaign-start} CLI command. */
 export async function handleCampaignStart(
@@ -58,45 +29,12 @@ export async function handleCampaignStart(
 ): Promise<void> {
   const cdpPort = options.cdpPort ?? DEFAULT_CDP_PORT;
 
-  // Reject conflicting options
-  if (options.personIds && options.personIdsFile) {
-    process.stderr.write(
-      "Use only one of --person-ids or --person-ids-file.\n",
-    );
-    process.exitCode = 1;
-    return;
-  }
-
-  // Parse person IDs from options
   let personIds: number[];
-  if (options.personIds) {
-    try {
-      personIds = parsePersonIds(options.personIds);
-    } catch (error) {
-      const message = errorMessage(error);
-      process.stderr.write(`${message}\n`);
-      process.exitCode = 1;
-      return;
-    }
-  } else if (options.personIdsFile) {
-    try {
-      personIds = readPersonIdsFile(options.personIdsFile);
-    } catch (error) {
-      const message = errorMessage(error);
-      process.stderr.write(`${message}\n`);
-      process.exitCode = 1;
-      return;
-    }
-  } else {
-    process.stderr.write(
-      "Either --person-ids or --person-ids-file is required.\n",
-    );
-    process.exitCode = 1;
-    return;
-  }
-
-  if (personIds.length === 0) {
-    process.stderr.write("No person IDs provided.\n");
+  try {
+    personIds = resolvePersonIds(options);
+  } catch (error) {
+    const message = errorMessage(error);
+    process.stderr.write(`${message}\n`);
     process.exitCode = 1;
     return;
   }
